0.5.0 - Initial Organization
The main goal of this 0.5 update is refactor the entire mission to be much more tidy and utilize proper script remoteExecs instead of the extremely hacky remoteExec["call"] non-sense I've been doing. This will require a significant change to decription.ext and massive amounts of rewrites of almost every script, but will lead to much better future work and enable easier porting.

For this initial commit, I have taken an initial stab at organizing the mission files into proper folders. The next step will be to define the locations of all of the scripts as simple strings to enable current compatibility. The step after that will be properly defining fnc's

Change:
- Organized all files into folders

Known Issues:
- Mission will not currently function due to script movement

Todo:
- Define server locations within init.sqf-called string list and update all instances with string list
- Define proper fnc type function calls