v0.5.2 - Auto-Zone Generation & QoL Improvements
This is a huge update under the hood for people who want to make ports of Overrun to new maps. I changed the ZoneArray master data set to be automatically generated based on square zone markers placed in the editor instead of "hard coded" into the init.sqf. This generated ZoneArray is then used to automatically create all of the zone triggers, removing the requirement to hand-place 30-50 triggers with very specific activation and deactivation code. The net effect of these changes brings the Overrun port time to about 2-3 hours instead of 10-12+ hours (depending on fat-fingering triggers).

This update also adds the ability for the arsenal-building script (initArsental.sqf) to be set only to load weapons and wearables from specific mods/CDLCs, meaning if you port it to, say, SOG: Prairie Fire, the players won't have access to vanilla Arma clothes or weapons. This is done via the item's "author" field, so look at the item's cfgWeapons or cfgEquipment field in the Config Viewer to find the correct author name to input into initArsenal.sqf.

After some multiplayer testing, I will be building a Porting Tutorial that covers how to bring Overrun to any map!

Lastly, this update includes some long overdue quality-of-life features, such as respawning with last selected loadout. See more in the changelog below

Major Changes:
- New ZoneArray generation script (initArray.sqf) searches for all Eden Editor placed markers of type "RECTANGLE" and builds the master data array based on the marker's name and size, including determining the maximum number of zombies. This is a change from the previous ZoneArray script which simply was hand coded in init.sqf.

- New system for creating the zone triggers (initTriggers.sqf) that automatically generates triggers from the ZoneArray and replaces the previously hand-placed triggers that initiate the zombie spawning and other zone-related functions.

- New system for the arsenal item filtering script (initArsenal.sqf) that allows map makers to filter out all clothing, backpacks, vests, hats, and weapons that are not related to a specific mod. This allows time-period specific variants of Overrun, such as SOG: Prairie Fire. See the instructions at the top of the initArsenal.sqf file.

- Added several quality of life features that have been requested. Firstly, all players now spawn with a full kit instead of "naked". Secondly, a script now saves the player's loadout when the arsenal is closed and respawns them with that same loadout instead of with nothing.

Minor Bugfixes:
- Corrected an issue with the deconTruck's respawn script that prevented the DECON addAction from running for some players.
- resetState.sqf and the save/load functions of saveState.sqf and init.sqf have been modified to accommodate new ZoneArray generation system.

Known Bugs/To Do:
- If a server already has a saved state for a scenario and the map is modified to add or remove zones, the scripts will have undesirable reactions due to different array sizes. Proposed fix: Add a check for the difference in saved values versus new number of zones and parse new default values into the added/removed zones.
- Players are still able to load previously saved loadouts that have unauthorized items (NVGs, etc). While this isn't an issue I am too worried about (who cares if you cheat in a coop game?), a half dozen or so people have reported it to me as a major flaw in the scenario design. Proposed fix: create a system that checks player's loadout against a list of naughty items and removes them (with a warning) if found
- Add more mini-objectives in the zones since people are getting bored of killing zombie bosses and finding little research boxes